[tps] 게시판 정보 등록/수정/삭제 기본 테스트 완료

+++ moka-core/moka-core-tps/src/main/java/jmnet/moka/core/tps/mvc/board/repository/BoardInfoRepository.java
@@ -4,6 +4,6 @@ import jmnet.moka.core.tps.mvc.board.entity.BoardInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface BoardInfoRepository extends JpaRepository<BoardInfo, Long>, JpaSpecificationExecutor<BoardInfo>, BoardInfoRepositorySupport {
+public interface BoardInfoRepository extends JpaRepository<BoardInfo, Integer>, JpaSpecificationExecutor<BoardInfo>, BoardInfoRepositorySupport {
 
 }
